refactor: pull out messages into separate file
