+++ commandline.c
@@ -5,7 +5,7 @@ void printPrompt() {
 }
 
 char* getInput() {
-    char buffer[1024];
+    char buffer[COMMANDLINE_BUFFER_SIZE];
     memset(buffer, '\0', sizeof(buffer));
     fgets(buffer, sizeof(buffer), stdin);
     int length = strlen(buffer);
@@ -15,10 +15,6 @@ char* getInput() {
     char* ret = malloc(strlen(buffer) * sizeof(char) + 1);
     strncpy(ret, buffer, strlen(buffer));
     return ret;
-}
-
-void printWelcome() {
-    printf("Welcome to EDF anonymizer\n");
 }
 
 void printMain(char* filename) {
@@ -34,14 +30,7 @@ void printMain(char* filename) {
     }
 }
 
-void printNotNumber(char* string) {
-    printf("ERROR: '%s' is not a number!  Please try again.\n", string);
-}
 
-void printUnknownOption(int option) {
-    printf("ERROR: Unknown option '%d'!  Please try again.\n", option);
-}
-
-void printNoOpenFile() {
-    printf("ERROR: No file currently open!  Please load a file first.");
+void printFilenameInput() {
+    printf("Please enter the filename to load.\n");
 }
