+++ commandline.h
@@ -5,7 +5,7 @@
 #include <string.h>
 #include <stdlib.h>
 
-// values for menu options and values
+// values for main menu options and values
 #define EXIT_STRING "Exit"
 #define EXIT_VAL 1
 #define LOAD_STRING "Load a .edf file"
@@ -16,6 +16,7 @@
 #define SAVE_VAL 4
 
 #define PROMPT_STRING "==> "
+#define COMMANDLINE_BUFFER_SIZE 1024
 
 /**
  * Gets input from the user.
@@ -28,27 +29,8 @@ char* getInput();
 void printPrompt();
 
 /**
- * Print a welcoming message.
- */
-void printWelcome();
-
-/**
  * Prints the main loop options.
  */
 void printMain(char* filename);
-
-/*
-* Error messages
-*/
-
-/**
- * Print a message indicating the input is not a valid option, and indicate which option was not valid.
- */
-void printUnknownOption(int option);
-
-/**
- * Print a message indicating the input is not a number.
- */
-void printNotNumber(char* string);
 
 #endif /* COMMANDLINE_H */
