+++ messages.c
@@ -1,0 +1,17 @@
+#include "messages.h"
+
+void printWelcome() {
+    printf("Welcome to EDF anonymizer\n");
+}
+
+void printNotNumber(char* string) {
+    printf("ERROR: '%s' is not a number!  Please try again.\n", string);
+}
+
+void printUnknownOption(int option) {
+    printf("ERROR: Unknown option '%d'!  Please try again.\n", option);
+}
+
+void printNoOpenFile() {
+    printf("ERROR: No file currently open!  Please load a file first.");
+}
